Refactor analytics supervisor setup
Instead of one worker to do refresh all google analytic tokens, we're
going to use a worker for each team, and a supervisor to supervise all the
 workers. We'll use gproc to dynamically generate names for the teams
